+++ broker/lib/index.js
@@ -3,6 +3,8 @@
 export { sopsDecrypt, sopsEncryptAtomic } from './sops.js';
 export { send, readBody, jsonError } from './http.js';
 export { aliyunRpcVersion, mergeAliyunQuery, ALIYUN_RPC_VERSION_BY_HOST } from './aliyun-rpc.js';
+export { resolveHostnameDoH, dohConnect, shouldSkipDoH, clearDoHCache } from './doh.js';
+export { defaultServiceTest, matchServiceTemplate, describeUpstreamStatus } from './service-test.js';
 export { defaultHealthBind, startLocalHealthServer, describeHealthBind } from './local-health.js';
 export { applyAdminVisibility, isAdminIdentity } from './admin-visibility.js';
 export { buildZip, computeCrc32 } from './zip.js';